Ergo/Sigmaparticle Helper project file updates: add password connect.py and deploy logic and import passwordFileEncryption and related functions add senderEIP3Secret arg to main function call, remove direct calls to os.getenv to connect.py sequence only, implement password argument call logic, replace print outputs with sys.stdout.write outputs, remove args like wallet_mnemonic, mnemonic_password, senderAddress in helper projects where they are not used, implement REST call instead of ergpy call due to some error in treeFromBox and implement def REST_getbox function in that helper project, implement REST call in treeToAddr helper project, SigmaParticleInterface.py updates: add any given passwords to function arguments and script call usages, check for normal or encrypted env file in SigmaParticle_updateKeyEnv function, add tries and sleep loop sequence to SigmaParticle_getTreeFromBox function, add ergopassword=, otherchainpassword= to checkBoxValue function and implement in script calls that it uses, Ergo/SigmaParticle/cpp/new_frame.cpp updates: check for existing swap / contract path before doing cp call to avoid accidently copying the entire basic_framework dir into the new dir, instead copy its contents if the contract / swap dir already exists Config Tools updates: add nodeAPIKey to TestnetErgo .env construction, not actually needed to be used yet but should be useful, might become optional main.py updates: SigmaParticle_box_to_addr, SigmaParticle_box_to_addr, GeneralizedENC_ResponderClaimSubroutine, checkBoxValue, main args endpoints with passwords responderInterface.py updates: implement password usages in function calls

+++ Ergo/SigmaParticle/cpp/new_frame.cpp
@@ -1,4 +1,6 @@
 #include <bits/stdc++.h>
+#include <iostream>
+#include <filesystem>
 #include <vector>
 using std::cout, std::string, std::vector;
 void makeNewContractFrame(string ContractName);
@@ -18,8 +20,18 @@ int main(int argc, char* argv[])
 
 void makeNewContractFrame(string ContractName)
 {
-	string command = "cp -r basic_framework " + ContractName;
-	system(command.c_str());	
+	std::string dirpath = ContractName;
+	if (std::filesystem::exists(dirpath) && std::filesystem::is_directory(dirpath))
+	{
+		string command = "cp -r basic_framework/* " + ContractName;
+		system(command.c_str());	
+	}
+	else
+	{
+		//if it doesnt exist yet copy and create, this will copy folder into existing folder if it existed
+		string command = "cp -r basic_framework " + ContractName;
+                system(command.c_str());
+	}
 }
 
 void echoContractName(string ContractName)
